put create/delete functions to common.py
