+++ heat/heat_tests.py
@@ -77,64 +77,10 @@ class HeatIntegrationTests(unittest.TestCase):
         https://mirantis.testrail.com/index.php?/cases/view/543347
         [Alexander Koryagin]
         """
-
         # Variables
-        # TODO :akoryagin: Be sure that this template file will be put on controller during test preparation
-        file_name = './Templates/stack_create_template.yaml'  # File with template for stack creation
-
-        new_stack_name = 'Test_{0}'.format(str(time.time())[0:10:])  # Like: 'Test_1449484927'
-        timeout_value = 10  # Timeout in minutes to wait for stack status change
-
-        def create_stack(heatclient, stack_name, template):
-            """ Create a stack from template and check STATUS == CREATE_COMPLETE
-                    :param heatclient: Heat API client connection point
-                    :param stack_name: Name of a new stack
-                    :param template:   Content of a template name
-                    :return uid: UID of stack
-            """
-            stack = heatclient.stacks.create(
-                stack_name=stack_name,
-                template=template,
-                parameters={})
-            uid = stack['stack']['id']
-
-            stack = heatclient.stacks.get(stack_id=uid).to_dict()
-            timeout = time.time() + 60 * timeout_value  # default: 10 minutes of timeout to change stack status
-            while stack['stack_status'] == 'CREATE_IN_PROGRESS':
-                stack = heatclient.stacks.get(stack_id=uid).to_dict()
-                if time.time() > timeout:
-                    break
-                else:
-                    time.sleep(5)
-
-            # Check that final status of a newly created stack is 'CREATE_COMPLETE'
-            self.assertEqual(
-                (stack['stack_status']), 'CREATE_COMPLETE',
-                msg='Stack failed to create: {}'.format(stack)
-            )
-            return uid
-
-        def delete_stack(heatclient, uid):
-            """ Delete stack and check STATUS == DELETE_COMPLETE
-                    :param heatclient: Heat API client connection point
-                    :param uid:        UID of stack
-            """
-            heatclient.stacks.delete(uid)
-
-            stack = heatclient.stacks.get(stack_id=uid).to_dict()
-            timeout = time.time() + 60 * timeout_value   # default: 10 minutes of timeout to change stack status
-            while stack['stack_status'] == 'DELETE_IN_PROGRESS':
-                stack = heatclient.stacks.get(stack_id=uid).to_dict()
-                if time.time() > timeout:
-                    break
-                else:
-                    time.sleep(5)
-
-            # Check that final status of a newly deleted stack is 'DELETE_COMPLETE'
-            self.assertEqual(
-                (stack['stack_status']), 'DELETE_COMPLETE',
-                msg='Stack fall to unknown status: {}'.format(stack)
-            )
+        # Be sure that this template file will be put on controller during test preparation
+        file_name = './Templates/stack_create_template.yaml'          # File with template for stack creation
+        new_stack_name = 'Test_{0}'.format(str(time.time())[0:10:])   # Like: 'Test_1449484927'
 
         # - 1 -
         # Read Heat stack-create template from file
@@ -146,12 +92,11 @@ class HeatIntegrationTests(unittest.TestCase):
 
         # - 2 -
         # Create new stack
-        uid_of_new_stack = create_stack(self.heat, new_stack_name, template_content)
+        uid_of_new_stack = common_functions.create_stack(self.heat, new_stack_name, template_content)
 
         # - 3 -
         # Delete created stack
-        delete_stack(self.heat, uid_of_new_stack)
-
+        common_functions.delete_stack(self.heat, uid_of_new_stack)
 
     def test_543337_HeatStackUpdate(self):
         """ This test case checks stack-update action.
